requirements.txt
streamlit>=1.30.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0
opencv-python>=4.8.0
PyWavelets>=1.4.0
